Add first pass type-specific * commands mechanism.
The type-specific commands are added to the end of the command list.
This affects their abbreviations, and it means they come at the end of
the *HELP output, which is no longer in remotely alphabetical order.

Should do a pass on this.

(*HELP does need to print commands in their actual order, so that it's
obvious how abbreviations will work.)