Fix status byte handling.
Correctly save status bytes in (hopefully...) all cases.

Document restrictions on packets with 0-byte payloads. They have no
status bytes, so the client has to assume no speculative responses
follow. Arguably a flaw in the serial encoding, but too late to fix now.
Easily documented and checked for in the server - the Beeb will now do a
BRK if it happens.

This restriction applies to all link types.